Move mock fixtures to separate file
